requirements.txt
Pillow
rembg
streamlit
numba
numpy